backend/requirements.txt
Django==6.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
